+++ src/main/java/org/classapp/backend/data/model/Role.java
@@ -2,7 +2,7 @@ package org.classapp.backend.data.model;
 
 public class Role {
 
-	private Integer id;
+	private Long id;
 	private String name;
 	private String description;
 	/**
@@ -10,7 +10,7 @@ public class Role {
 	 * @param name
 	 * @param description
 	 */
-	public Role(Integer id, String name, String description) {
+	public Role(Long id, String name, String description) {
 		this.id = id;
 		this.name = name;
 		this.description = description;
@@ -23,13 +23,13 @@ public class Role {
 	/**
 	 * @return the id
 	 */
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 	/**
 	 * @param id the id to set
 	 */
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 	/**
